Refactor: type 코드 분리
